+++ src/survey/survey.module.ts
@@ -4,8 +4,8 @@ import { Survey } from './entities/survey.entity';
 import { Question } from './entities/question.entity';
 import { Option } from './entities/option.entity';
 import { Answer } from './entities/answer.entity';
-import { SurveyService } from './survey.service';
-import { SurveyController } from './survey.controller';
+import { SurveyService } from './services/survey.service';
+import { SurveyController } from './controller/survey.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Survey, Question, Option, Answer])],
